requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
Pillow==8.4.0
tqdm==4.62.3